db: Extract hashColumns() to be reusable
