+++ master/buildbot/db/changesources.py
@@ -22,6 +22,7 @@ from twisted.internet import defer
 
 from buildbot.db import NULL
 from buildbot.db import base
+from buildbot.util.sautils import hash_columns
 from buildbot.warnings import warn_deprecated
 
 
@@ -57,7 +58,7 @@ class ChangeSourceModel:
 class ChangeSourcesConnectorComponent(base.DBConnectorComponent):
     def findChangeSourceId(self, name):
         tbl = self.db.model.changesources
-        name_hash = self.hashColumns(name)
+        name_hash = hash_columns(name)
         return self.findSomethingId(
             tbl=tbl,
             whereclause=(tbl.c.name_hash == name_hash),
